+++ src/juego/interfaces/Atacable.java
@@ -3,7 +3,7 @@ package juego.interfaces;
 public interface Atacable {
 
 	public void regenerar();
-	public void daniar(int danio);
+	public void daniar(float danio);
 	public boolean vidaAgotada();
 	public void deshabilitar();
 	public void afectadoPorRadiacion();
